requirements.txt
fastapi
google-cloud-bigquery
firebase-admin
uvicorn